Bootstrap basic Next.js blog
`npx create-next-app` + a couple of style tweaks.

I chose to use tailwind for styling because it's become the most common
tool in the industry, at least for prototyping, from what I can see.
I'm particularly a fan (at all). And went for Typescript because it is a
great productivity and quality aid.